Fixed osx build without templates.

+++ plugin/collation_dictionary/character_sets.cc
@@ -59,7 +59,7 @@ bool CharacterSetsTool::Generator::populate()
       push(tmp_cs->comment);
 
       /* MAXLEN */
-      push((int64_t) tmp_cs->mbmaxlen);
+      push(tmp_cs->mbmaxlen);
 
       cs++;
